refactor: move tests to eval module
