+++ src/eval.rs
@@ -89,3 +89,42 @@ pub fn eval(expr: &SExpr) -> Result<Value, EvalError> {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn assert_evaluates_to(expr: &str, value: i64) {
+        let result = eval(&expr.parse().unwrap()).unwrap();
+        assert_eq!(result, Value::Integer(value));
+    }
+
+    #[test]
+    fn test_add() {
+        assert_evaluates_to("(+ 2 3)", 5);
+        assert_evaluates_to("(+ 4 5)", 9);
+    }
+
+    #[test]
+    fn test_mul() {
+        assert_evaluates_to("(* 2 3)", 6);
+        assert_evaluates_to("(* 4 5)", 20);
+    }
+
+    #[test]
+    fn test_atom() {
+        assert_evaluates_to("3", 3);
+    }
+
+    #[test]
+    fn test_nested() {
+        assert_evaluates_to("(+ 1 (* 2 3))", 7);
+        assert_evaluates_to("(+ (* 1 2) (* 3 (+ 4 5)))", 29);
+    }
+
+    #[test]
+    fn test_eval_error() {
+        let result = eval(&"(+ 2 *)".parse().unwrap());
+        assert!(result.is_err())
+    }
+}
