+++ src/main.rs
@@ -30,43 +30,3 @@ fn main() {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::eval::EvalError;
-
-    fn assert_evaluates_to(expr: &str, value: i64) {
-        let result = parse_eval(expr).unwrap();
-        assert_eq!(result, Value::Integer(value));
-    }
-
-    #[test]
-    fn test_add() {
-        assert_evaluates_to("(+ 2 3)", 5);
-        assert_evaluates_to("(+ 4 5)", 9);
-    }
-
-    #[test]
-    fn test_mul() {
-        assert_evaluates_to("(* 2 3)", 6);
-        assert_evaluates_to("(* 4 5)", 20);
-    }
-
-    #[test]
-    fn test_atom() {
-        assert_evaluates_to("3", 3);
-    }
-
-    #[test]
-    fn test_nested() {
-        assert_evaluates_to("(+ 1 (* 2 3))", 7);
-        assert_evaluates_to("(+ (* 1 2) (* 3 (+ 4 5)))", 29);
-    }
-
-    #[test]
-    fn test_eval_error() {
-        let err = parse_eval("(+ 2 *)").unwrap_err();
-        assert!(err.is::<EvalError>());
-    }
-}
